Add files via upload
Topic: College inquiry chat bot system
• Description: A chatbot is an artificial person, animal, or other creature
which holds conversations with humans. The motivation for this project
comes from websites and the internet where often the information provided
is by machines, not humans. The aim of this project is to create and
implement a chatbot system that can offer human-like answers to
academic queries related to admissions, fee structure, scholarship details,
timetable, document requirements, etc.